requirements.txt
absl-py==2.3.1
alembic==1.14.0
APScheduler==3.11.0
asgiref==3.9.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
immutabledict==4.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.14.1
llvmlite==0.44.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numba==0.61.2
numpy==2.2.6
ordered-set==4.1.0
ortools==9.14.6206
packaging==24.2
pandas==2.3.3
pillow==11.3.0
protobuf==6.31.1
psycopg2-binary==2.9.10
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
rich==13.9.4
scipy==1.16.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
tqdm==4.67.1
typer==0.16.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
vortexcrypt==1.0.7
Werkzeug==3.1.3
whitenoise==6.11.0
wrapt==1.17.0
